Small cleanup (#496)
* Drop a RowPanel member re-declaration prior to usage

Flagged by PyCharm.  It looks like this snuck in with c665f21f107e.

* Fix iteration over nested panels in a Dashboard

This appears to be an obvious typo (after PyCharm flagged `row_panel` as
unused), but it looks like this is only used by `auto_panel_ids()`.  I wasn't
able to generate a difference in output before and after this change when adding
2 `SingleStat` panels to a `RowPanel`, and in turn adding that to the dashboard.
The changed code definitely runs in this case, so IDK what is ultimately making
the changes to the nested panels prior to this.